% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{addJacobian}
\alias{addJacobian}
\title{Return jacobian function}
\usage{
addJacobian(fn, includeJacobian)
}
\arguments{
\item{fn}{Jacobian for the specified transformation.}

\item{includeJacobian}{Whether to actually use a jacobian transformation.}
}
\value{
A function, either the jacobian originally passed as fn, or a function that only returns 0.
}
\description{
Return jacobian function
}
\details{
Add ADoverload to user passed functions to ensure that they work within the RTMB package.
}
