% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{addPrior}
\alias{addPrior}
\title{Return prior function}
\usage{
addPrior(fn, jac = NULL, includeJacobian = FALSE)
}
\arguments{
\item{fn}{Prior distribution as a function.}

\item{jac}{Jacobian transformation as a function (Default NULL).}

\item{includeJacobian}{Logical to include the jacobian or not (Default FALSE).}
}
\value{
A function, either the prior originally passed as fn, or a function that only returns 0.
}
\description{
Return prior function
}
\details{
Add ADoverload to user passed functions to ensure that they work within the RTMB package.
}
