% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMixture.r
\name{basicMixtureModel}
\alias{basicMixtureModel}
\title{Basic Mixture Model}
\usage{
basicMixtureModel(
  x,
  K = NULL,
  mu = NULL,
  sigma = NULL,
  prob = NULL,
  mu_fixed = NULL,
  sigma_fixed = NULL,
  wgts = NULL,
  names = NULL,
  control = list()
)
}
\arguments{
\item{x}{data as a vector.}

\item{K}{number of components, default = NULL and ignored if names are given.}

\item{mu}{starting values for the mean (default = NULL).}

\item{sigma}{starting values for the standard deviation (default = NULL).}

\item{prob}{starting values for mixture proportions (default = NULL).}

\item{mu_fixed}{named vector with values to hold fixed (default = NULL).}

\item{sigma_fixed}{named vector with values to hold fixed (default = NULL).}

\item{wgts}{mixture weights associated with each observation x (default = NULL).}

\item{names}{Names of each component in the mixture model (default = NULL) Must provide names or K.}

\item{control}{List for specifying fitting details of EM algorithm (see details).}
}
\value{
List of model estimates.
}
\description{
Basic Gaussian mixture model fit using the EM algorithm.
}
\details{
The control list input
\itemize{
\item \code{maxiters} Maximum iterations before failing (default = 1000).
\item \code{tolerance} Maximum difference between log-likelihood between each iteration to set convergence (default = 1e-8).
\item \code{relativeDifference} TRUE then convergence is taken as relative difference between the log-likelihood on each iteration, or FALSE it is total difference (default = FALSE).
\item \code{verbose} TRUE then print more details upon fitting, or FALSE fit quietly unless there is an error.
}
}
\examples{
\dontrun{
p <- c(0.2, 0.7, 0.1)
mu <- c(10, 25, 40)
sigma <- c(1.25, 2, 3.3)
x <- rnorm(200, mu[sample(3, 200, prob = p, replace = TRUE)], sigma[sample(3, 200, prob = p, replace = TRUE)])
fit <- basicMixtureModel(x, K = 3)

est <- NULL
est2 <- NULL
for( i in 1:500 ){
   N <- c(2000, 1000)
   nsmp <- 25
   p1 <- c(0.3, 0.4, 0.3)
   p2 <- c(0.15, 0.65, 0.2)
   ptrue <- (p1*N[1] + p2*N[2])/sum(N)
   id1 <- sample(3, nsmp, prob = p1, replace = TRUE)
   id2 <- sample(3, nsmp, prob = p2, replace = TRUE)
   x1 <- rnorm(nsmp, c(25, 45, 65)[id1], c(3, 3.7, 2.5)[id1])
   x2 <- rnorm(nsmp, c(25, 45, 65)[id2], c(3, 3.7, 2.5)[id2])
   wgts <- c(rep(N[1]/nsmp, nsmp), rep(N[2]/nsmp, nsmp))
   fit <- basicMixtureModel(x=c(x1, x2), K = 3, prob = c(0.2, 0.4, 0.4), mu = c(20, 40, 60), wgts = wgts)
   fit2 <- basicMixtureModel(x=c(x1, x2), K = 3, prob = c(0.2, 0.4, 0.4), mu = c(20, 40, 60))
   est <- rbind(est, data.frame(par = c("p1", "p2", "p3"), diff = fit$proportion[order(fit$mu)] - ptrue, method = "weighted"))
   est <- rbind(est, data.frame(par = c("p1", "p2", "p3"), diff = fit2$proportion[order(fit2$mu)] - ptrue, method = "unweighted"))
}
boxplot(diff~method+par, data = est)
abline(h = 0, col='red')
}
}
