% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.r
\name{calcPostProb}
\alias{calcPostProb}
\title{Calculate the Posterior Probability of Mixture Proportions}
\usage{
calcPostProb(x, mu, sigma, prob, wgts = NULL)
}
\arguments{
\item{x}{vector of observations (e.g. fish lengths).}

\item{mu}{vector of means for each component (e.g. species).}

\item{sigma}{vector of standard deviations for each componenet (e.g. species).}

\item{prob}{matrix of mixture weights for each observation and component. Often constant but may vary due to covariates.}

\item{wgts}{vector of likelihood weights for each observation.}
}
\value{
list of posterior probabilities and log likelihood of mixture component.
}
\description{
Used in the Expectation-Maximization Algorithm for fitting the joint mixture model.
}
\details{
Computes the posterior probability of each x belonging to component k for a given mean (\code{mu}), standard deviation (\code{sigma}), and mixture weight (\code{prob}).
Allows for likelihood weights to be passed to compute the log-likelihood of the mixture component.
}
