% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.r
\name{calcProportions}
\alias{calcProportions}
\title{Calculate the Mixture Proportions for Prediction in model:}
\usage{
calcProportions(alpha, alpha_jackchinook, p_adultchinook, Xprop, K0, K)
}
\arguments{
\item{alpha}{vector of linear predictors for each species proportions.}

\item{alpha_jackchinook}{vector of linear predictors for jack Chinook as a proportion of all Chinook.}

\item{Xprop}{is a list of design matrices with ncol = \code{length(alpha)}, and the final one ncol = \code{length(alpha_jackchinook)}.}

\item{K0}{Number of non-Chinook species}

\item{K}{Number of total categories (species + Chinook types).}

\item{pAdultChinook}{proportion of adult Chinook that are small or large. Defaults to fixed = 1 if adults are just one group.}
}
\value{
list of posterior probabilities and log likelihood of mixture component.
}
\description{
Used in the Expectation-Maximization Algorithm for fitting the joint mixture model.
}
