% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processInputs.r
\name{checkDate}
\alias{checkDate}
\title{Check value is a in correct Date format}
\usage{
checkDate(e_date)
}
\arguments{
\item{e_date}{Value to check.}
}
\value{
A date in Y-m-d format.
}
\description{
Check value is a in correct Date format
}
