% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesianMixture.r
\name{compute_param_values}
\alias{compute_param_values}
\title{Compute posterior parameter values}
\usage{
compute_param_values(self, post, ad_obj)
}
\arguments{
\item{self}{The R6 \code{speciesCompModel} that is set up for a model run.}

\item{post}{Matrix of simualted parameter values that are directly estimated.}

\item{ad_obj}{AD model object that was used to fit that we can then generate true values from.}
}
\value{
matrix of simulated values transformed to predicted values.
}
\description{
Compute posterior parameter values
}
