% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.r
\name{dbinomc}
\alias{dbinomc}
\title{Binomial distribution allowing for continuous size.}
\usage{
dbinomc(x, size, prob, log = FALSE)
}
\arguments{
\item{x}{Number of successes}

\item{size}{Number of trials (can be continuous).}

\item{prob}{Probability of success.}

\item{log}{Logical, return log density (TRUE) or on real scale (FALSE).}
}
\value{
Density of binomial distribution of x.
}
\description{
Binomial distribution allowing for continuous size.
}
