% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.r
\name{dlogitnorm_interval}
\alias{dlogitnorm_interval}
\title{Logitnormal interval distribution}
\usage{
dlogitnorm_interval(x, mean = 0, sd = 1, lower = 0, upper = 1, log = FALSE)
}
\arguments{
\item{x}{value on the real scale.}

\item{mean}{Mean on logit interval scale.}

\item{sd}{Standard deviation on logit interval scale.}

\item{log}{Logical, return log density (TRUE) or on real scale (FALSE).}
}
\value{
Density of log normal distribution for value x.
}
\description{
Logitnormal interval distribution
}
