% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.r
\name{dlognorm}
\alias{dlognorm}
\title{Lognormal distribution (mean and sd on log scale)}
\usage{
dlognorm(x, mean, sd, log = FALSE)
}
\arguments{
\item{x}{value on the real scale.}

\item{mean}{Mean on log scale of the log normal distribution.}

\item{sd}{Standard deviation on log scale of the log normal distribution.}

\item{log}{Logical, return log density (TRUE) or on real scale (FALSE).}
}
\value{
Density of log normal distribution for value x.
}
\description{
Lognormal distribution (mean and sd on log scale)
}
