% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.r
\name{dlognorm_real}
\alias{dlognorm_real}
\title{Lognormal distribution (mean and sd on real scale)}
\usage{
dlognorm_real(x, mean_real, sd_real, log = FALSE)
}
\arguments{
\item{x}{value on the real scale}

\item{mean_real}{Mean on real scale of the log normal distribution.}

\item{sd_real}{Standard deviation on real scale of the log normal distribution.}

\item{log}{Logical, return log density (TRUE) or on real scale (FALSE).}
}
\value{
Density of log normal distribution for value x.
}
\description{
Lognormal distribution (mean and sd on real scale)
}
