% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{estimate_daily_proportions}
\alias{estimate_daily_proportions}
\title{Daily proportion of salmon}
\usage{
estimate_daily_proportions(prop, pred_df, species)
}
\arguments{
\item{prop}{Proportion of fish at level of provided covariates, estimated by species composition model.}

\item{pred_df}{dataframe with information needed to aggregate prop to daily estimates.}

\item{species}{vector of standard deviations for each componenet (e.g. species).}
}
\value{
Matrix of daily proportions for each species.
}
\description{
Combine proportions based on estimates for covariates to the daily scale.
}
