% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{expitM}
\alias{expitM}
\title{Inverse Multivariate Logit Transform}
\usage{
expitM(x)
}
\arguments{
\item{x}{vector to transform}
}
\value{
transformed value adding an extra dimension such that sums to 1.
}
\description{
Inverse Multivariate Logit Transform
}
\details{
Transform from (-Inf, Inf) -> (0, 1).
}
