% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{extractControls}
\alias{extractControls}
\title{Extract controls}
\usage{
extractControls(controlValue, defaultValue)
}
\arguments{
\item{controlValue}{Value passed to function}

\item{defualtValue}{Value set as a default. If \code{controlValue} is NULL returns this value.}
}
\value{
either control value if it is not NULL, or default value.
}
\description{
Extract controls
}
