% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{extractParams}
\alias{extractParams}
\title{Extract parameter values}
\usage{
extractParams(
  self,
  init_values = NULL,
  fixed_values = NULL,
  default_values,
  name,
  transform = I,
  ...
)
}
\arguments{
\item{self}{R6 speciesCompModel object.}

\item{init_values}{initial values as a named vector.}

\item{fixed_values}{values that are to be held fixed in the model.}

\item{default_values}{values to fill in if init_values are not provided and are not fixed.}

\item{name}{vector of the names of the vector values that need to be returned.}

\item{transform}{function to transform the values to the scale needed for estimation.}

\item{...}{additional values to pass to the transformation.}
}
\value{
No return, R6 object is updated for both \verb{$params_init} and \verb{$params_fixed}
}
\description{
Extract parameter values
}
