% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{findGlobalMode}
\alias{findGlobalMode}
\title{Find global mode of a vector.}
\usage{
findGlobalMode(x, xrange = NULL, ...)
}
\arguments{
\item{x}{vector of values to find the modes of.}

\item{xrange}{range of values to look for the global mode in.}

\item{...}{additional values to pass to the \code{density} function.}
}
\value{
Returns scalar.
}
\description{
Find the global mode using a kernel density approximation.
}
\details{
Returns largest value found by density approximation.
}
