% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{findModes}
\alias{findModes}
\title{Find modes of a vector.}
\usage{
findModes(x, ...)
}
\arguments{
\item{x}{vector of values to find the modes of.}

\item{...}{additional values to pass to the \code{density} function.}
}
\value{
Return modes found, or NA if none are found.
}
\description{
Find the local modes using a kernel density approximation.
}
\details{
Any time the density increases and then decreases, a mode is identified.
}
