% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processInputs.r
\name{fitChinookLengths}
\alias{fitChinookLengths}
\title{Fit Chinook Lengths}
\usage{
fitChinookLengths(
  x,
  chinook_names,
  mu_chin,
  sigma_chin,
  proportions_chin,
  mu_fixed = NULL,
  sigma_fixed = NULL
)
}
\arguments{
\item{x}{Vector of Chinook length values.}

\item{mu_chin}{Named vector with mean chinook values assumed.}

\item{proportions_chin}{Named vector of expected proportion of each Chinook age/size class.}

\item{mu_fixed}{Named vector of mean lengths that were provided by the user to hold fixed.}

\item{sigma_fixed}{Named vector of standard deviation of lengths that were provided by the user to hold fixed.}

\item{sigma}{Named vector with standard deviation of chinook values assumed.}
}
\value{
List with mean, standard deviation, and proportions for Chinook.
}
\description{
Fit Chinook Lengths
}
