% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{fn_null}
\alias{fn_null}
\title{Call functions with null values.}
\usage{
fn_null(f, ...)
}
\arguments{
\item{f}{function to do transformation.}

\item{...}{additional function information.}
}
\value{
Function that returns NULL if input is NULL.
}
\description{
Call functions with null values.
}
\details{
Generalizes R functions to deal with NULL values.
}
