% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{histplot}
\alias{histplot}
\title{Weighted histogram for plotting length data}
\usage{
histplot(
  x,
  wgts,
  range = c(35, 120),
  delta = 2,
  ylim = NA,
  xlim = NA,
  xlab = NA,
  plot = TRUE,
  ...
)
}
\arguments{
\item{x}{Vector of values to pass to histogram.}

\item{wgts}{Weights of each value.}

\item{range}{Range that the histogram plot will cover.}

\item{delta}{The step size to bin by.}

\item{ylim}{ylimit to be used.}

\item{xlim}{xlimit to be used.}

\item{xlab}{xlabel to plot with.}

\item{plot}{logical to either plot the histogram or not.}

\item{...}{Additional arguments for \code{barplot} function call.}
}
\value{
list of range to plot under (\verb{$range}), the binned values (\verb{$lengths}), and histplot \verb{$density}.
}
\description{
Weighted histogram for plotting length data
}
\details{
Taken from \code{plotrix} package for the weighted histogram, reworked for specific purposes.
}
