% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{ilogitInterval}
\alias{ilogitInterval}
\title{Inverse interval transform logit}
\usage{
ilogitInterval(x, lower, upper)
}
\arguments{
\item{x}{Value to transform}

\item{lower}{Lower bound of the value to be transformed}

\item{upper}{Upper bound of the value to be transformed}
}
\value{
value on original scale (lower - upper).
}
\description{
Inverse interval transform logit
}
\details{
inverse logit transformation for a value centered at zero to become (0,1). Then scaled to lower and upper.
}
