% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{joinPars}
\alias{joinPars}
\title{Join fixed and estiamted parameter values into a single vector.}
\usage{
joinPars(est = NULL, fixed = NULL, species)
}
\arguments{
\item{est}{values that might be estimated.}

\item{fixed}{values that might be fixed}

\item{species}{names of the values to join in correct order.}
}
\value{
vector of combined fixed and estimated parameter values.
}
\description{
Join fixed and estiamted parameter values into a single vector.
}
