% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{log_null}
\alias{log_null}
\title{log function with null values.}
\usage{
log_null(x)
}
\arguments{
\item{x}{value to log.}
}
\value{
log(x) or NULL if is.null(x).
}
\description{
log function with null values.
}
