% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{log_sum_exp}
\alias{log_sum_exp}
\title{Log Sum Exponential}
\usage{
log_sum_exp(x)
}
\arguments{
\item{x}{vector on the log scale to sum.}
}
\value{
stable version of \code{sum(x)}.
}
\description{
Log Sum Exponential
}
\details{
Find the max value, subtract that on the log scale and exponentiate and sum. Then return back on log scale.
}
