% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{logDetJac_logitInterval}
\alias{logDetJac_logitInterval}
\title{Log Determinant Jacobian of logit interval transformation.}
\usage{
logDetJac_logitInterval(x, lower, upper)
}
\arguments{
\item{x}{Value to transform}

\item{lower}{Lower bound of the value to be transformed}

\item{upper}{Upper bound of the value to be transformed}
}
\value{
log determinant of jacobian.
}
\description{
Log Determinant Jacobian of logit interval transformation.
}
\details{
Compute the jacobian of the transformation to go from (lower, upper) -> (-Inf, Inf).
}
