% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{logitInterval}
\alias{logitInterval}
\title{Interval transform logit}
\usage{
logitInterval(x, lower, upper)
}
\arguments{
\item{x}{Value to transform}

\item{lower}{Lower bound of the value to be transformed}

\item{upper}{Upper bound of the value to be transformed}
}
\value{
transformed value
}
\description{
Interval transform logit
}
\details{
Transform from (lower, upper) -> (-Inf, Inf).
}
