% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{logitM}
\alias{logitM}
\title{Multivariate Logit Transform}
\usage{
logitM(x)
}
\arguments{
\item{x}{vector to transform}
}
\value{
transformed value of reduced dimension.
}
\description{
Multivariate Logit Transform
}
\details{
Transform from (0, 1) -> (-Inf, Inf).
}
