% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesianMixture.r
\name{make_negll}
\alias{make_negll}
\title{Joint likelihood function for fitting with RTMB:}
\usage{
make_negll(self)
}
\arguments{
\item{self}{The R6 \code{speciesCompModel} that is set up for a model run.}
}
\value{
Function that is the negative log likelihood.
}
\description{
Joint likelihood function for fitting with RTMB:
}
\details{
Negative log likelihood function for fitting species composition model.
}
