% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{mission_2022}
\alias{mission_2022}
\title{Mission Site Data Year 2022}
\format{
\subsection{\code{mission_2022}}{

A list:
\describe{
\item{sonarCounts}{Sonar counts file data frame.}
\item{sonarLengths}{Measured sonar lengths data frame.}
\item{albion}{Albion test fishery catch data.}
\item{whonnock}{Whonnock test fishery catch data.}
\item{brownsvillebar}{Brownesville Bar test fishery catch data.}
\item{salmonPassage}{Daily salmon passage estimate, which includes large resident.}
\item{testFisheryLength}{Test fishery length data.}
}
}
}
\usage{
mission_2022
}
\description{
Data from Mission hydroacoustic site for the entire 2022 season.
}
\keyword{datasets}
