% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot_beam_spreading}
\alias{plot_beam_spreading}
\title{Plot Beam Spreading Trend}
\usage{
plot_beam_spreading(self)
}
\arguments{
\item{self}{R6 speciesCompModel object.}
}
\description{
Plot Beam Spreading Trend
}
\details{
Plot the change in expected species length against sonar range (m).
}
