% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot_mix}
\alias{plot_mix}
\title{Plot weighted histogram along with estimated proportions.}
\usage{
plot_mix(self, day = 1, ...)
}
\arguments{
\item{self}{R6 speciesCompModel object.}

\item{day}{Day to plot (default = 1). Day 1 is most recent day.}

\item{...}{Additional options to pass to plot function.}

\item{include_proportion_labels}{Include labels in plot (default = FALSE).}
}
\value{
list of range to plot under (\verb{$range}), the binned values (\verb{$lengths}), and histplot \verb{$density}.
}
\description{
Plot weighted histogram along with estimated proportions.
}
\details{
Taken from \code{plotrix} package for the weighted histogram, reworked for specific purposes. This function
is called directly from \verb{$plotMix} within the R6 object.
}
