% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot_prior}
\alias{plot_prior}
\title{Plot prior distribution against fitted posterior mode.}
\usage{
plot_prior(self, parameter)
}
\arguments{
\item{self}{R6 speciesCompModel object.}

\item{parameter}{Which parameter to plot, on the transformed scale (e.g. "log_qinv").}
}
\description{
Plot prior distribution against fitted posterior mode.
}
\details{
Makes a plot of the histogram to help identify issues with the prior and fitted values.
}
