% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot_test_fishery_lengths}
\alias{plot_test_fishery_lengths}
\title{Plot Test Fishery Lengths}
\usage{
plot_test_fishery_lengths(self, test_fishery_lengths = NULL, ndays = 6)
}
\arguments{
\item{self}{R6 speciesCompModel object.}

\item{test_fishery_lengths}{Data frame of test fishery lengths with Date, FL.cm, and Species.}

\item{ndays}{Number of days to combine for test fishing length data.}
}
\description{
Plot Test Fishery Lengths
}
\details{
Plot histogram and density plots assumed for the species lengths data.
}
