% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.r
\name{prior_assess}
\alias{prior_assess}
\title{Generate points along the prior distribution provided by the user.}
\usage{
prior_assess(self, parameter)
}
\arguments{
\item{self}{R6 object for sonar species composition analysis.}

\item{parameter}{Name of the parameter to generate the prior for.}
}
\description{
Generate points along the prior distribution provided by the user.
}
