% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processInputs.r
\name{process_albion_catch}
\alias{process_albion_catch}
\title{Process Albion Catch}
\usage{
process_albion_catch(self, albion_catch)
}
\arguments{
\item{self}{R6 speciesCompModel Object.}

\item{albion_catch}{Data frame with columns "CRPT_DTT", "NET_START_OUT", "NET_FULL_OUT", "NET_START_IN", "NET_FULL_IN", and "NET_CONFIG".}
}
\value{
No object returned, appends test fishery catch to \code{self}, \code{self$test_fishery_catch[["albion"]]}.
}
\description{
Process Albion Catch
}
