% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{process_minutes}
\alias{process_minutes}
\title{Process Minutes}
\usage{
process_minutes(time)
}
\arguments{
\item{time}{Character in H:M:S format.}
}
\value{
Time in hours.
}
\description{
Process Minutes
}
