% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processInputs.r
\name{process_mission_catch}
\alias{process_mission_catch}
\title{Process PSC Catch}
\usage{
process_mission_catch(
  self,
  test_fishery_counts,
  name = "whonnock",
  tangled = TRUE
)
}
\arguments{
\item{self}{R6 speciesCompModel Object.}

\item{test_fishery_counts}{Data frame in standard format for Whonnock catch.}

\item{name}{Test fishery name (default = "whonnock").}

\item{tangle}{Logical whether to include or exclude tangled counts.}
}
\value{
No object returned, appends test fishery catch to \code{self$test_fishery_catch[[name]]}
}
\description{
Process PSC Catch
}
\details{
input data frame should have columns start_out, full_out, start_in, full_in along with TRIP_DTT, and net_length. The other
required columns are generally \verb{Sockeye Adult Gilled} and \verb{Sockeye Adult Tangled}.
}
