% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processInputs.r
\name{process_mission_salmon_passage}
\alias{process_mission_salmon_passage}
\title{Process Mission Salmon Passage}
\usage{
process_mission_salmon_passage(self, salmon_passage_table)
}
\arguments{
\item{self}{R6 speciesCompModel Object.}

\item{salmon_passage_table}{Data frame in standard format for Mission count based on Right and Left Bank.}
}
\value{
No object returned, appends salmon counts \code{self$salmon_counts}
}
\description{
Process Mission Salmon Passage
}
