% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processQualarkInputs.r
\name{process_qualark_catch}
\alias{process_qualark_catch}
\title{Process Qualark Catch}
\usage{
process_qualark_catch(self, test_fishery_counts)
}
\arguments{
\item{self}{R6 speciesCompModel Object.}

\item{test_fishery_counts}{Data frame in standard format for Whonnock catch.}
}
\value{
No object returned, appends test fishery catch to \code{self$test_fishery_catch[[name]]}
}
\description{
Process Qualark Catch
}
\details{
input data frame should have columns start_out, full_out, start_in, full_in along with TRIP_DTT, and net_length. The other
required columns are generally \verb{Sockeye Adult Gilled} and \verb{Sockeye Adult Tangled}.
}
