% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processQualarkInputs.r
\name{process_qualark_lengths}
\alias{process_qualark_lengths}
\title{Process Qualark Sonar Lengths}
\usage{
process_qualark_lengths(self, sonar_counts, sonar_lengths, dropN = 2)
}
\arguments{
\item{self}{R6 speciesCompModel Object.}

\item{sonar_counts}{Sonar Lengths in standard Mission format.}

\item{sonar_lengths}{Sonar Lengths in standard Mission Format.}

\item{dropN}{Minimum sample size in a stratum.}
}
\value{
No object returned, appends sonar_lengths to self.
}
\description{
Process Qualark Sonar Lengths
}
