% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processInputs.r
\name{process_whonnock_catch}
\alias{process_whonnock_catch}
\title{Process Whonnock Catch}
\usage{
process_whonnock_catch(self, test_fishery_counts, tangled = TRUE)
}
\arguments{
\item{self}{R6 speciesCompModel Object.}

\item{test_fishery_counts}{Data frame in standard format for Whonnock catch.}

\item{tangle}{Logical whether to include or exclude tangled counts.}
}
\value{
No object returned, appends test fishery catch to \code{self$test_fishery_catch[["whonnock"]]}
}
\description{
Process Whonnock Catch
}
\details{
input data frame should have columns start_out, full_out, start_in, full_in along with TRIP_DTT, and net_length. The other
required columns are generally \verb{Sockeye Adult Gilled} and \verb{Sockeye Adult Tangled}.
}
