% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{rbind_lazy}
\alias{rbind_lazy}
\title{Join estimated parameters:}
\usage{
rbind_lazy(pars1 = NULL, pars2 = NULL)
}
\arguments{
\item{pars1}{First object to rbind}

\item{pars2}{Second object to rbind}
}
\value{
Two objects joined with NA in missing columns.
}
\description{
Join estimated parameters:
}
