% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{reList}
\alias{reList}
\title{Put parameters into original list format.}
\usage{
reList(parsList, parsVec)
}
\arguments{
\item{parsList}{original pars list to push values to.}

\item{parsVec}{vector of par values to push into list.}
}
\value{
updated parsList with values from parsVec.
}
\description{
Put parameters into original list format.
}
