% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesianMixture.r
\name{run_full_model}
\alias{run_full_model}
\title{Fit joint species composition model as a Bayesian model.}
\usage{
run_full_model(self, bayesian = TRUE)
}
\arguments{
\item{self}{The R6 \code{speciesCompModel} that is set up for a model run.}
}
\value{
List of model estimates (\verb{$params_estimated}), convergence details (\verb{$convergence}),
estimate of daily species passage (\verb{$N_daily}), and daily proportions (\verb{$p_daily}).
}
\description{
Expectation-Maximization algorithm for fitting the joint hydroacoustic lengths and test fishery model.
}
\details{
Techincal details for this function are provided as a vignette. Computes posterior probabilities of lengths to species
given current values of the parameters. Those probabilities are passed to the objective function and then the objective function is maximized to
find the new parameters. This process is repeated until the log-likelihood becomes stable and the parameter estimates are then returned along with
being being stored in the R6 object $params_estimated.
}
