% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMixture.r
\name{runEM}
\alias{runEM}
\title{Run EM Algorithm}
\usage{
runEM(EM, control)
}
\arguments{
\item{EM}{Taped RTMB model object. See details.}

\item{control}{List to control convergence criteria of the algorithm. See details.}
}
\value{
List of model estimates (\verb{$params_estimated}), convergence details (\verb{$convergence}),
estimate of daily species passage (\verb{$N_daily}), and daily proportions (\verb{$p_daily}).
}
\description{
Iterate through the Expectation-Maximization algorithm until convergence criteria are met.
}
\details{
The EM object is built by other functions and takes parameter estimates, returns updated parameter estimates and the log-likelihood.
Typically we assume that object is taped and built with RTMB but that is not necessary, as long as it does the inner optimization of the objective function.
The control list input is
\itemize{
\item \code{maxiters} Maximum iterations before failing (default = 1000).
\item \code{tolerance} Maximum difference between log-likelihood between each iteration to set convergence (default = 1e-8).
\item \code{relativeDifference} TRUE then convergence is taken as relative difference between the log-likelihood on each iteration, or FALSE it is total difference (default = FALSE).
\item \code{verbose} TRUE then print more details upon fitting, or FALSE fit quietly unless there is an error.
}
}
