% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processInputs.r
\name{set_daily_data}
\alias{set_daily_data}
\title{Set Daily Data}
\usage{
set_daily_data(self)
}
\arguments{
\item{self}{R6 speciesCompModel Object.}
}
\value{
No object returned, appends a data list with all the daily data needed for the model \code{self$data_list}
}
\description{
Set Daily Data
}
\details{
Process daily data for the model to prepare for analysis.
}
