% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processInputs.r
\name{set_length_adjustment}
\alias{set_length_adjustment}
\title{Set Length Adjustment}
\usage{
set_length_adjustment(self, formula = NULL)
}
\arguments{
\item{self}{R6 speciesCompModel Object.}

\item{formula}{Formula for how to model beam spreading.}
}
\value{
No object returned, appends a data list with all the daily data needed for the model \code{self$data_list$X_length}
}
\description{
Set Length Adjustment
}
\details{
Default formula is \code{~ beamWidth.cm} will use mean correction of L.cm - beta_0 - beta_1 * beamWidth.cm. Other suggestion
is to use SonarBin. Any columns available within \code{self$data_list$length_data} is available here.
}
