% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processInputs.r
\name{set_model_parameters}
\alias{set_model_parameters}
\title{Set Model Parameters}
\usage{
set_model_parameters(
  self,
  fixed_parameters = c("mu", "sigma", "proportion_jackchinook"),
  fixed_values = list(),
  initial_values = list(),
  delta_mu_bounds = NULL,
  delta_sd_bounds = NULL
)
}
\arguments{
\item{self}{R6 speciesCompModel Object.}

\item{fixed_parameters}{Character vector with entire parameters to hold fixed.}

\item{fixed_values}{List with names equal to parameters and named vectors providing fixed values.}

\item{initial_values}{List with names equal to parameters and named vectors providing initial values.}

\item{delta_mu_bounds}{Matrix (first column lower bound, second upper bound) or vector (lower bound, upper bound) (Defaults to (-1,1)).}

\item{delta_sd_bounds}{Matrix (first column lower bound, second upper bound) or vector (lower bound, upper bound) (Defaults to (-0.5,0.5)).}
}
\value{
No object returned, appends \code{self$params_fixed} and \code{self$params_init}
}
\description{
Set Model Parameters
}
\details{
Set initial values and fixed values for running the species composition model. Intended to be updated before \code{self$fitModel}.
}
