% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processInputs.r
\name{set_model_proportions}
\alias{set_model_proportions}
\title{Set Model Proportions}
\usage{
set_model_proportions(self, formula = list())
}
\arguments{
\item{self}{R6 speciesCompModel Object.}

\item{formula}{List of formulas for proportions of different species.}
}
\value{
No object returned, appends a data list with all the daily data needed for the model \code{self$data_list}
}
\description{
Set Model Proportions
}
\details{
If formula in list \code{list("all" = ~factor(day))} then this is applied to all species except Chinook. Default formula is currently
\code{~-1+stratum:factor(day)}.
}
