% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processInputs.r
\name{set_species_lengths}
\alias{set_species_lengths}
\title{Set Species Length}
\usage{
set_species_lengths(
  self,
  mu = NULL,
  sigma = NULL,
  proportions_chinook = NULL,
  test_fishery_lengths = NULL,
  ndays = 7
)
}
\arguments{
\item{self}{R6 speciesCompModel Object.}

\item{mu}{Named vector of lengths to assume in the model e.g. \code{c("largeresident" = 35)}.}

\item{sigma}{Named vector of lengths to assume in the model e.g. \code{c("largeresident" = 2.5)}.}

\item{proportions_chinook}{Proportion of Chinook that are different age/size classes ("jackchinook", "adultchinook").}

\item{test_fishery_lengths}{Data frame with columns ("FL.cm", "Date", "Species").}

\item{ndays}{Number of days to include for estimating expected mean and variance of lengths for each species.}
}
\value{
No object returned, appends a intial values to model object \code{self}, \code{self$default_parameters}.
}
\description{
Set Species Length
}
\details{
If a mu value is not provided and test_fishery_lengths are, then we will try and estimate mean and variance based on the
number of days provided (default = 7). If not enough catches are available, or test fishery data is not provided, then default values are used.
For Chinook, if proportions_chinook is not provided, we attempt to estimate this through a mixture model.
}
