% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{speciesColours}
\alias{speciesColours}
\title{Species colours}
\usage{
speciesColours(species)
}
\arguments{
\item{species}{Vector of species names}
}
\value{
Vector of colours matching the species names.
}
\description{
Species colours
}
