% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{speciesLabels}
\alias{speciesLabels}
\title{Species labels}
\usage{
speciesLabels(species)
}
\arguments{
\item{species}{Vector of species names}
}
\value{
Vector of names for display.
}
\description{
Labels to plot species names.
}
