% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{estimate_daily_salmon}
\alias{estimate_daily_salmon}
\title{Estimate daily salmon}
\usage{
estimate_daily_salmon(p_daily, total_salmon, species)
}
\arguments{
\item{p_daily}{Proportion of fish each day, estimated by species composition model.}

\item{total_salmon}{Daily salmon escapement counted by the hydroacoustic program.}

\item{species}{vector of standard deviations for each componenet (e.g. species).}
}
\value{
Matrix for each species, excluding small resident fish, with an additional column added that
combines all Chinook salmon.
}
\description{
Combine proportions and total salmon to estimate daily species escapement
}
\details{
If small resident are included, they are removed when estimating total daily salmon.
Large resident fish are included (> 35 cm for updated Mission protocols).
}
