% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot_test_fishery}
\alias{plot_test_fishery}
\title{Plot Pearson residuals of test fishery model (CPUE - Nq)}
\usage{
plot_test_fishery(self)
}
\arguments{
\item{self}{R6 speciesCompModel object.}
}
\description{
Plot Pearson residuals of test fishery model (CPUE - Nq)
}
\details{
Plot the Pearson residuals for the test fishery catch, displayed as CPUE - Nq. x-axis labels are species - test fishery - net type.
}
