% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Simulate Lengths and Test Fishery Data:}
\usage{
simulate(
  N = c(1000, 5000, 2000),
  nLengths = 20,
  sonarBins = c(1, 2, 3),
  sonarBanks = c(1, 2),
  mu = c(38, 58, 70),
  sigma = c(4, 3.5, 7),
  beta = c(0, 0),
  sigmaObs = 5.5,
  q = c(1, 1),
  control = list()
)
}
\arguments{
\item{N}{Vector containing the total number of fish for each species.}

\item{nLengths}{Total number of lengths measurements to take.}

\item{sonarBins}{vector of sonar bins with which counts and lengths are taken. Assumed 1,2,3.}

\item{sonarBanks}{vector of sonar banks that counts and lengths are taken. Assumed 1 left and 2 right bank.}

\item{mu}{Vector of mean length for each species.}

\item{sigma}{Vector of standard deviation of length for each species.}

\item{beta}{Vector of correction terms for intercept and beam spreading effect.}

\item{sigmaObs}{Observational error associated with each measured length.0}

\item{q}{Vector of relative catchability of length equal to length(N) - 1, relative to the last fish.}

\item{control}{List that allows for controlling additional simulation parameters. See 'details' for a description of all the options.}
}
\description{
Function to simulate sonar lengths and test fishery data to test hydroacoustic species composition model.
}
\examples{
datalist <- simulate(N = c(1000, 50000, 20000), nLengths = 20, 
                     sonarBins = c(1,2,3), sonarBanks = c(1,2), mu = c(38, 58, 70), sigma = c(4,3.5,7),
                     beta = c(1.5, 0.8), sigmaObs)

}
