% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processInputs.r
\name{speciesCompModel}
\alias{speciesCompModel}
\title{Base class for species composition estimation methods}
\description{
An R6 base class that is used to step through all the different setups and steps of estimating species composition on the Fraser River (Qualark and Mission) using Hydroacoustics and the test fishery.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{species_info}}{List with information about species naming and ordering.}

\item{\code{est_date}}{Date for the day of analysis.}

\item{\code{ndays}}{Number of days to join.}

\item{\code{data_list}}{List of all data used for analysis for the ndays of interest.}

\item{\code{data_info}}{List of information specific to site and analysis (e.g. test fishery information).}

\item{\code{fit_info}}{List of optimiziation and fitting information for fitting the model.}

\item{\code{sonar_lengths}}{Dataframe of sonar lengths passed for entire data processed.}

\item{\code{test_fishery_catch}}{List of test fishery catch for each fishery.}

\item{\code{salmon_counts}}{Dataframe daily salmon counts estimate.}

\item{\code{params_estimated}}{List of all estimated parameters after fitting model.}

\item{\code{params_fixed}}{List all values to hold fixed when fitting the model.}

\item{\code{params_init}}{List all initial values to use when fitting the model.}

\item{\code{sim_data}}{List of daily simulated data to match data_list.}

\item{\code{default_parameters}}{List of default parameter values to use when inits are not provided.}

\item{\code{prior_distributions}}{List of prior distributions for all parameters.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SpeciesCompModel-new}{\code{speciesCompModel$new()}}
\item \href{#method-SpeciesCompModel-setSpecies}{\code{speciesCompModel$setSpecies()}}
\item \href{#method-SpeciesCompModel-setDate}{\code{speciesCompModel$setDate()}}
\item \href{#method-SpeciesCompModel-controlOptimization}{\code{speciesCompModel$controlOptimization()}}
\item \href{#method-SpeciesCompModel-processData}{\code{speciesCompModel$processData()}}
\item \href{#method-SpeciesCompModel-setSpeciesLengths}{\code{speciesCompModel$setSpeciesLengths()}}
\item \href{#method-SpeciesCompModel-setModelParameters}{\code{speciesCompModel$setModelParameters()}}
\item \href{#method-SpeciesCompModel-fitModel}{\code{speciesCompModel$fitModel()}}
\item \href{#method-SpeciesCompModel-simulate}{\code{speciesCompModel$simulate()}}
\item \href{#method-SpeciesCompModel-setPriors}{\code{speciesCompModel$setPriors()}}
\item \href{#method-SpeciesCompModel-plotMix}{\code{speciesCompModel$plotMix()}}
\item \href{#method-SpeciesCompModel-plotTestFishery}{\code{speciesCompModel$plotTestFishery()}}
\item \href{#method-SpeciesCompModel-plotBeamSpreading}{\code{speciesCompModel$plotBeamSpreading()}}
\item \href{#method-SpeciesCompModel-plotTestFisheryLengths}{\code{speciesCompModel$plotTestFisheryLengths()}}
\item \href{#method-SpeciesCompModel-priorCheck}{\code{speciesCompModel$priorCheck()}}
\item \href{#method-SpeciesCompModel-clone}{\code{speciesCompModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpeciesCompModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-SpeciesCompModel-new}{}}}
\subsection{Method \code{new()}}{
Initialize the R6 object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{speciesCompModel$new(
  species = c("jackchinook", "sockeye", "adultchinook"),
  date = NULL,
  ndays = 1,
  site = NULL,
  roll_angle = NULL,
  feasible_lengths = c(10, 120)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{species}}{vector of species names  e.g. c("largeresident", "jackChinook", "sockeye", "adultchinook").}

\item{\code{date}}{Date that you will estimate. Place holder but does not need to be set.}

\item{\code{ndays}}{Number of days to combine, defaults to 1.}

\item{\code{site}}{Choose between "Mission" or "Qualark", roll angle is set be default based on site.}

\item{\code{roll_angle}}{The roll angle that the sonar is set to (default = NULL).}

\item{\code{feasible_lengths}}{Lower and upper bound of feasible lengths that can be measured on the sonar to remove infeasible (default = c(10, 120)).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpeciesCompModel-setSpecies"></a>}}
\if{latex}{\out{\hypertarget{method-SpeciesCompModel-setSpecies}{}}}
\subsection{Method \code{setSpecies()}}{
Update the R6 object with species names to fit in a model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{speciesCompModel$setSpecies(species = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{species}}{Character vector with options are "smallresident", "largeresident", "pink", "sockeye", "coho", "chum", "jackchinook", "adultchinook", "smalladultchinook", "largeadultchinook"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpeciesCompModel-setDate"></a>}}
\if{latex}{\out{\hypertarget{method-SpeciesCompModel-setDate}{}}}
\subsection{Method \code{setDate()}}{
Update the estimation date and number of days to use.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{speciesCompModel$setDate(date = NULL, ndays = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{date}}{Date to change the analysis date to.}

\item{\code{ndays}}{Number of days to combine.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpeciesCompModel-controlOptimization"></a>}}
\if{latex}{\out{\hypertarget{method-SpeciesCompModel-controlOptimization}{}}}
\subsection{Method \code{controlOptimization()}}{
Control optimization of the EM algorithm.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{speciesCompModel$controlOptimization(control = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{control}}{List with values \verb{$tolerance} what difference is okay (default 1e-8), \verb{$maxiters} maximum iterations of EM algorithm (default 1000), \verb{$relative_difference} true or false to choose relative vs absolute difference on each iteration of algorithm, \verb{$verbose} true or false to print details of EM algorithm.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpeciesCompModel-processData"></a>}}
\if{latex}{\out{\hypertarget{method-SpeciesCompModel-processData}{}}}
\subsection{Method \code{processData()}}{
Function to process input data for analysis. Does the main data processing and adds a few columns needed elsewhere for the entire dataset added to make it fast and easy to do different days.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{speciesCompModel$processData(
  sonar_counts,
  sonar_lengths,
  test_fishery_counts = NULL,
  include_tangled = TRUE,
  dropN = 2,
  salmon_passage_table = NULL,
  feasible_lengths = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sonar_counts}}{Data frame of sonar counts in standard format.}

\item{\code{sonar_lengths}}{Data frame of sonar lengths in standard format.}

\item{\code{test_fishery_counts}}{List of dataframes named after the test fishery (e.g. 'whonnock' or 'albion')}

\item{\code{include_tangled}}{Logical whether to use the tangled fish in the gilnet.}

\item{\code{dropN}}{Minimum number of observations to keep length measurements per hour/bin/bank etc. if > dropN (default 2) but not always necessary.}

\item{\code{salmon_passage_table}}{Data frame that has the total number of salmon predicted each day at the site.}

\item{\code{feasible_lengths}}{Minimum and maximum lengths that are feasible. Set by default but can be changed here.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpeciesCompModel-setSpeciesLengths"></a>}}
\if{latex}{\out{\hypertarget{method-SpeciesCompModel-setSpeciesLengths}{}}}
\subsection{Method \code{setSpeciesLengths()}}{
Set species length parameter information
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{speciesCompModel$setSpeciesLengths(
  mu = NULL,
  sigma = NULL,
  proportion_adultchinook = NULL,
  test_fishery_lengths = NULL,
  ndays = 6,
  plot = FALSE,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mu}}{Named vector the user wants to set as default values \code{c("largeresident" = 38)}.}

\item{\code{sigma}}{Named vector the user wants to set as default values e.g. \code{c("largeresident" = 2.5)}.}

\item{\code{proportion_adultchinook}}{Named vector that sets default proportions of Chinook e.g. \code{c("jackchinook" = 0.05, "adultchinook" = 0.95)}.}

\item{\code{test_fishery_lengths}}{Data frame that contains test fishery lengths for use of parameter estimation default values.}

\item{\code{ndays}}{Number of prior days to use from test fishery lengths (default = 6).}

\item{\code{plot}}{Logical to plot the estimated mean and variance set by the function.}

\item{\code{verbose}}{Logical to print warning messages or not.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpeciesCompModel-setModelParameters"></a>}}
\if{latex}{\out{\hypertarget{method-SpeciesCompModel-setModelParameters}{}}}
\subsection{Method \code{setModelParameters()}}{
Set the model parameters before fitting with EM algorithm, including which values to fix and which to fit.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{speciesCompModel$setModelParameters(
  fixed_parameters = c("mu", "sigma", "proportion_jackchinook"),
  fixed_values = list(),
  initial_values = list(),
  formula_proportions = list(),
  formula_lengths = ~beamWidth.cm,
  date = NULL,
  ndays = NULL,
  delta_mu_bounds = NULL,
  expansion_formula = NULL,
  test_fishery_spp = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fixed_parameters}}{Which parameters to hold fixed (default \code{c("mu", "sigma", "proportion_jackchinook")}).}

\item{\code{fixed_values}}{List provided by the user to hold particular values fixed, if nothing provided but parameters are held fixed then default values will be used.}

\item{\code{initial_values}}{List provided by the user to improve fitting time for estimated parameters, default are used if not provided.}

\item{\code{formula_proportions}}{list of formulas to use for alpha for each species, if \code{list("all" = ~ factor(day))}, then each species is given the same formula, except jack chinook are treated separately.}

\item{\code{formula_lengths}}{Formula for how to deal with beam spreading (default = \code{~beamWidth.cm},}

\item{\code{date}}{Date (optional) to change analysis date.}

\item{\code{ndays}}{Number of days to combine (optional).}

\item{\code{delta_mu_bounds}}{Matrix or vector to schoose how much the mean length of each species can vary. Matrix must have nrows of number of species. Otherwise a vector is length 2 and declares lower and upper for all.}

\item{\code{expansion_formula}}{Formula to set the test fishery expansion line relationships (defualt = ~ species:fishery).}

\item{\code{test_fishery_spp}}{Vector of joined species and test fishery names to used e.g. ("sockeye_whonnock", "adultchinook_whonnock", "chinook_albion").}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpeciesCompModel-fitModel"></a>}}
\if{latex}{\out{\hypertarget{method-SpeciesCompModel-fitModel}{}}}
\subsection{Method \code{fitModel()}}{
EM algorithm for fitting the joint species composition model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{speciesCompModel$fitModel(control = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{control}}{List containing \code{controlOptimization} arguments as well as '$include_test_fishery' - logical to include or exclude test fishery catch information in the model, '$adjust_lengths' logical include beam spreading adjusment, and '$test_fishery_weights' the model weights for the test fishery.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
List of parameter estimates, which is also stored in the \code{params_estimated} field of the R6 object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpeciesCompModel-simulate"></a>}}
\if{latex}{\out{\hypertarget{method-SpeciesCompModel-simulate}{}}}
\subsection{Method \code{simulate()}}{
Simulate the model for the counts on the 'est_date' and 'ndays' with parameter values equal to 'params_estimated' within the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{speciesCompModel$simulate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpeciesCompModel-setPriors"></a>}}
\if{latex}{\out{\hypertarget{method-SpeciesCompModel-setPriors}{}}}
\subsection{Method \code{setPriors()}}{
Define Prior distributions (or penalty functions) for fitting the joint species composition model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{speciesCompModel$setPriors(priors = list(), includeJacobian = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{priors}}{List of prior distributions. Defaults to \code{\(...){0}}.}

\item{\code{includeJacobian}}{Logical of whether or not to include jacobian transformation.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpeciesCompModel-plotMix"></a>}}
\if{latex}{\out{\hypertarget{method-SpeciesCompModel-plotMix}{}}}
\subsection{Method \code{plotMix()}}{
Plot the fitted mixture model against the historgram of the length data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{speciesCompModel$plotMix(day = 1, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{day}}{to plot (day = 1 is \code{est_date}) (default = 1).}

\item{\code{...}}{Additional plot arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpeciesCompModel-plotTestFishery"></a>}}
\if{latex}{\out{\hypertarget{method-SpeciesCompModel-plotTestFishery}{}}}
\subsection{Method \code{plotTestFishery()}}{
Plot Pearson residuals for the test fishery data (CPUE - Nq) against each species, test fishery, and net type.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{speciesCompModel$plotTestFishery(includePrior = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{includePrior}}{Should the prior distribution be plotted alongside the estimates.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpeciesCompModel-plotBeamSpreading"></a>}}
\if{latex}{\out{\hypertarget{method-SpeciesCompModel-plotBeamSpreading}{}}}
\subsection{Method \code{plotBeamSpreading()}}{
Plot estimated beam spreading effects against sonar range.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{speciesCompModel$plotBeamSpreading(includePrior = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{includePrior}}{Should the prior distribution be plotted alongside the estimates.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpeciesCompModel-plotTestFisheryLengths"></a>}}
\if{latex}{\out{\hypertarget{method-SpeciesCompModel-plotTestFisheryLengths}{}}}
\subsection{Method \code{plotTestFisheryLengths()}}{
Plot test fishery lengths against the assumed mean and standard deviation to be used in the model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{speciesCompModel$plotTestFisheryLengths(test_fishery_lengths = NULL, ndays = 6)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{test_fishery_lengths}}{Length data frame that contains Date, FL.cm, and Species.}

\item{\code{ndays}}{Number of days to join for plotting test fishery data.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpeciesCompModel-priorCheck"></a>}}
\if{latex}{\out{\hypertarget{method-SpeciesCompModel-priorCheck}{}}}
\subsection{Method \code{priorCheck()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{speciesCompModel$priorCheck(parameter)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpeciesCompModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SpeciesCompModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{speciesCompModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
